Moved replacetags to new file, rename triggermessage function
